Extract CompositeStorage interface.
We need this so that the CompositeStorageImpl can listen to other CompositeStorage interfaces in the future.
