+++ refinedstorage2-network-api/src/test/java/com/refinedmods/refinedstorage2/api/network/test/StorageChannelTypes.java
@@ -7,6 +7,7 @@ import com.refinedmods.refinedstorage2.api.storage.channel.StorageChannelImpl;
 import com.refinedmods.refinedstorage2.api.storage.channel.StorageChannelType;
 import com.refinedmods.refinedstorage2.api.storage.channel.StorageTracker;
 import com.refinedmods.refinedstorage2.api.storage.composite.CompositeStorage;
+import com.refinedmods.refinedstorage2.api.storage.composite.CompositeStorageImpl;
 
 import java.util.Collections;
 import java.util.List;
@@ -24,7 +25,7 @@ public final class StorageChannelTypes {
 
         @Override
         public CompositeStorage<String> createCompositeStorage(List<Storage<String>> sources) {
-            return new CompositeStorage<>(sources, new ResourceListImpl<>());
+            return new CompositeStorageImpl<>(sources, new ResourceListImpl<>());
         }
     };
